refactor(task-graph): extract StreamProcessor from TaskRunner
Moves the 185-line executeStreamingTask body into a new internal class.
StreamProcessor.run takes (input, ctx, deps) where deps bundles facade
state pulled at call time (registry, resourceScope, inputStreams) and
facade methods bound to the facade instance (onProgress, own).

Method-arg passing keeps StreamProcessor stateless beyond its task reference;
the facade re-reads its own mutable per-run state (registry, resourceScope,
inputStreams) at each call instead of having StreamProcessor cache stale
values.

executeStreamingTask was never overridden by any subclass, so removing it
from TaskRunner has zero subclass impact.

Pure refactor: no behavior change, no public API change. 1,669 integration
tests pass unchanged including the streaming-specific suites.
